# cwp-28/cwp-28/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
host: localhost:3000
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /agents/{id}:
    x-swagger-router-controller: agentscontroller
    get:
      description: Agents
      operationId: get
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResultResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ResultResponse"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "string"
    post:
      description: persons
      operationId: update
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Any description"
        required: true
        schema:
          $ref: "#/definitions/NewAgent"
      responses:
        405:
          description: "Invalid input"
    delete:
      description: persons
      operationId: deleteF
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "integer"
      responses:
        405:
          description: "Invalid input"
  /foods/{id}:
    x-swagger-router-controller: foodscontroller
    get:
      description: Foods
      operationId: get
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResultResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ResultResponse"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "string"
    post:
      description: persons
      operationId: update
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Any description"
        required: true
        schema:
          $ref: "#/definitions/NewFood"
      responses:
        405:
          description: "Invalid input"
    delete:
      description: persons
      operationId: deleteF
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "integer"
      responses:
        405:
          description: "Invalid input"
  /agents/ingestions/{id}:
    x-swagger-router-controller: ingestionscontroller
    get:
      description: ingestions
      operationId: get
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResultResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ResultResponse"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "string"
    post:
      description: persons
      operationId: update
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Any description"
        required: true
        schema:
          $ref: "#/definitions/NewIngestion"
      responses:
        405:
          description: "Invalid input"
    delete:
      description: persons
      operationId: deleteF
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "integer"
      responses:
        405:
          description: "Invalid input"
  /persons/{id}:
    x-swagger-router-controller: personscontroller
    get:
      description: persons
      operationId: get
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResultResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ResultResponse"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "string"
    post:
      description: persons
      operationId: update
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Any description"
        required: true
        schema:
          $ref: "#/definitions/NewPerson"
      responses:
        405:
          description: "Invalid input"
    delete:
      summary: "Person"
      description: persons
      operationId: deleteF
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id"
        required: true
        type: "integer"
      responses:
        405:
          description: "Invalid input"
...
definitions:
  NewPerson:
    type: "object"
    properties:
      name:
        type: "string"
      address:
        type: "string"
      isAlive:
        type: "boolean"
    xml:
      name: "Person"
...
  NewFood:
    type: "object"
    properties:
      name:
        type: "string"
      calories:
        type: "integer"
        format: "int32"
    xml:
      name: "Food"
...
  NewAgent:
    type: "object"
    properties:
      callsign:
        type: "string"
    xml:
      name: "Agent"
...
  NewIngestion:
    type: "object"
    properties:
      when:
        type: "string"
      foodId:
        type: "integer"
        format: "int32"
      personId:
        type: "integer"
        format: "int32"
      reportedById:
        type: "integer"
        format: "int32"
    xml:
      name: "Ingestion"
...
  ResultResponse:
    properties:
      result:
        type: string
...